♻️ [Refactor]: create-study.service 이름 변경
